TASK - 2: enhance ux/ui - implement the auth flow - apply vuex & cookies logic - update controller to prevent xsrf and sql injection - add vue unti test

+++ src/controllers.php
@@ -4,6 +4,7 @@
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
+
 $app->extend('twig', function($twig, $app) {
     $twig->addGlobal('user', $app['session']->get('user'));
 
@@ -19,12 +20,84 @@ $app->match('/{path}', function () use ($app) {
 });
 
 
+$app->get('/api/requestAuth', function () use ($app) {
+    $xsrfToken = md5(uniqid());
+    $app['session']->set('xsrfToken', $xsrfToken);
+
+    return json_encode(array('token' => $xsrfToken));
+});
+
+$app->match('/api/login', function (Request $request) use ($app) {
+    $username = $app->escape($request->get('username'));
+    $password = $app->escape($request->get('password'));
+
+    $xsrfToken = $app['session']->get('xsrfToken');
+    $requestToken = $request->headers->get('Authorization');
+
+    if (empty($xsrfToken) || empty($requestToken) || $xsrfToken !== $requestToken) {
+        $response = new Response(json_encode(array('msg' => 'Request unauthorized')));
+        return $response->setStatusCode(401);
+    }
+
+    if ($username) {
+        $sql = "SELECT * FROM users WHERE username = '$username' and password = '$password'";
+        $user = $app['db']->fetchAssociative($sql);
+
+        if ($user){
+            $app['session']->set('user', $user);
+            return json_encode(
+                array(
+                    'id'        => $user['id'],
+                    'username'  => $user['username'],
+                )
+            );
+        }
+    }
+
+    $response = new Response(json_encode(array('msg' => 'Invalid username or password')));
+    return $response->setStatusCode(500);
+});
+
+
+$app->get('/api/logout', function () use ($app) {
+    $app['session']->set('user', null);
+    $app['session']->set('xsrfToken', null);
+    return json_encode(array('success' => true));
+});
+
+$app->match('/api/list', function (Request $request) use ($app) {
+    if (null === $user = $app['session']->get('user')) {
+        return $app->redirect('/login');
+    }
+
+    $xsrfToken = $app['session']->get('xsrfToken');
+    $requestToken = $request->headers->get('Authorization');
+
+    if (empty($xsrfToken) || empty($requestToken) || $xsrfToken !== $requestToken) {
+        $response = new Response(json_encode(array('msg' => 'Request unauthorized')));
+        return $response->setStatusCode(401);
+    }
+    $sql = "SELECT * FROM todos";
+    $todos = $app['db']->fetchAllAssociative($sql);
+
+    return $app->json($todos);
+});
+
+
 $app->get('/api/todo/{id}', function ($id, Request $request) use ($app) {
     if (null === $user = $app['session']->get('user')) {
         return $app->redirect('/login');
     }
 
     $contentType = $request->headers->get('Content-Type');
+
+    $xsrfToken = $app['session']->get('xsrfToken');
+    $requestToken = $request->headers->get('Authorization');
+
+    if (empty($xsrfToken) || empty($requestToken) || $xsrfToken !== $requestToken) {
+        $response = new Response(json_encode(array('msg' => 'Request unauthorized')));
+        return $response->setStatusCode(401);
+    }
 
     if ($id){
         $sql = "SELECT * FROM todos WHERE id = '$id'";
@@ -47,69 +120,41 @@ $app->get('/api/todo/{id}', function ($id, Request $request) use ($app) {
 })
     ->value('id', null);
 
-$app->match('/api/list', function () use ($app) {
-    if (null === $user = $app['session']->get('user')) {
-        return json_encode(array('authenticated' => false));
-    }
-
-    $sql = "SELECT * FROM todos";
-    $todos = $app['db']->fetchAllAssociative($sql);
-
-    return $app->json($todos);
-});
-
-
-
-$app->match('/login', function (Request $request) use ($app) {
-    $username = $request->get('username');
-    $password = $request->get('password');
-
-    if ($username) {
-        $sql = "SELECT * FROM users WHERE username = '$username' and password = '$password'";
-//        var_dump($app['db']); die;
-        $user = $app['db']->fetchAssociative($sql);
-
-        if ($user){
-            $app['session']->set('user', $user);
-            return $app->redirect('/todo');
-        }
-    }
-
-    return $app['twig']->render('login.html', array());
-});
-
-
-$app->get('/logout', function () use ($app) {
-    $app['session']->set('user', null);
-    return $app->redirect('/');
-});
-
-
-
-
-
-$app->post('/todo/add', function (Request $request) use ($app) {
+$app->post('/api/todo/add', function (Request $request) use ($app) {
     if (null === $user = $app['session']->get('user')) {
         return $app->redirect('/login');
     }
 
-    $user_id = $user['id'];
-    $description = $request->get('description');
-    $contentType = $request->headers->get('Content-Type');
+    $xsrfToken = $app['session']->get('xsrfToken');
+    $requestToken = $request->headers->get('Authorization');
 
+    if (empty($xsrfToken) || empty($requestToken) || $xsrfToken !== $requestToken) {
+        $response = new Response(json_encode(array('msg' => 'Request unauthorized')));
+        return $response->setStatusCode(401);
+    }
+
+    $user_id = $user['id'];
+    $description = $app->escape($request->get('description'));
+    $contentType = $request->headers->get('Content-Type');
     $sql = "INSERT INTO todos (user_id, description) VALUES ('$user_id', '$description')";
     $app['db']->executeUpdate($sql);
 
     if (strpos($contentType, 'application/json') === false) {
         return $app->redirect('/todo');
     } else {
-//        return json_encode(array('success' => true));
-        return $app->json(['response' => 'success']);
+       return json_encode(array('success' => true));
     }
 });
 
 
-$app->match('/todo/delete/{id}', function (Request $request, $id) use ($app) {
+$app->match('/api/todo/delete/{id}', function (Request $request, $id) use ($app) {
+    $xsrfToken = $app['session']->get('xsrfToken');
+    $requestToken = $request->headers->get('Authorization');
+
+    if (empty($xsrfToken) || empty($requestToken) || $xsrfToken !== $requestToken) {
+        $response = new Response(json_encode(array('msg' => 'Request unauthorized')));
+        return $response->setStatusCode(401);
+    }
 
     $sql = "DELETE FROM todos WHERE id = '$id'";
     $app['db']->executeUpdate($sql);
@@ -124,6 +169,13 @@ $app->match('/todo/delete/{id}', function (Request $request, $id) use ($app) {
 
 
 $app->match('/api/todo/complete/{id}', function (Request $request, $id) use ($app) {
+    $xsrfToken = $app['session']->get('xsrfToken');
+    $requestToken = $request->headers->get('Authorization');
+
+    if (empty($xsrfToken) || empty($requestToken) || $xsrfToken !== $requestToken) {
+        $response = new Response(json_encode(array('msg' => 'Request unauthorized')));
+        return $response->setStatusCode(401);
+    }
 
     $sql = "UPDATE todos SET completed = 1 WHERE id = '$id'";
     $app['db']->executeUpdate($sql);
